fix: Fixed and improved route53 module for getting consoleURL of clusters
